# Switch to OpenTelemetry tracing library