--- a/go.cap
+++ b/go.cap
@@ -1,10 +1,8 @@
 github.com/onflow/cadence ()
 
 github.com/davecgh/go-spew/spew (file)
-github.com/onflow/cadence/runtime/interpreter ()
 github.com/onflow/cadence/runtime/parser (file)
 github.com/onflow/cadence/runtime/pretty (runtime)
-github.com/opentracing/opentracing-go (network)
 github.com/stretchr/testify/assert (runtime, file, network)
 github.com/zeebo/blake3/internal/consts (file)
 github.com/klauspost/cpuid/v2 (runtime, file)