--- a/runtime/interpreter/interpreter_tracing_test.go
+++ b/runtime/interpreter/interpreter_tracing_test.go
@@ -23,8 +23,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/opentracing/opentracing-go"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/attribute"
 
 	"github.com/onflow/cadence/runtime/common"
 	"github.com/onflow/cadence/runtime/interpreter"
@@ -43,7 +43,7 @@ func setupInterpreterWithTracingCallBack(
 			func(inter *interpreter.Interpreter,
 				operationName string,
 				duration time.Duration,
-				logs []opentracing.LogRecord) {
+				attrs []attribute.KeyValue) {
 				tracingCallback(operationName)
 			},
 		),