Issue: https://github.com/onflow/flow-go/issues/2796

## Description
This diff switches [deprecated](https://github.com/opentracing/specification/issues/163) OpenTracing-based library with standardized OpenTelemetry API.  This is also needed since we are migrating Flow production tracing to OTEL Collector that prefers native OTEL formats and libraries.

OpenTelemetry says the following in their docs:
> There is no hard requirement to migrate existing application code to OpenTelemetry. However, we do recommend migrating from any OpenTracing instrumentation libraries – libraries used to instrument web frameworks, HTTP clients, database clients, etc. – to their OpenTelemetry equivalents. This will improve support, as many OpenTracing libraries will be retired and may no longer be updated.

PS. While here, regenerated `go.cap` though I have no idea what i'm doing there.
PPS. This commit will break flow-go interfaces, so it needs to be coupled with a similar diff there.  (Not sure if the "Breaking change" label should be used for this or this is only for user-visible changes.)
______

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels