@turbolent 2022-07-18
@SaveTheRbtz Could you please benchmark locally and see if this change has a significant impact?