@simonhf 2022-07-18
> @SaveTheRbtz Could you please benchmark locally and see if this change has a significant impact?

@turbolent FYI I think there are two schools of thought about this:
1. It doesn't matter how fast or slow open tracing is because we can just tweak the sampling rate appropriately, e.g. let's say the new API is twice as slow, so we'd just halve the sampling rate from every 16th TX to every 32nd TX... no harm done?
2. Use a faster way to collect the actual timing info (the part which is speed critical), and then do all the other processing asynchronously (the part where it doesn't matter how long it takes because we have lots of spare CPU).

I did some timing of the old Jaeger instrumentation here [1] and discovered that surprisingly it's even slower than using zerolog just for the stats collection part which presumably doesn't actually send off any data to the Jaeger server.

FYI I also suggest using a different form of instrumentation which only records the minimum info -- e.g. timestamp and context -- during the time critical code, into a simple memory buffer. Then at the end of the time critical code, the recorded info can be processed asynchronously and even sent to multiple places concurrently, e.g. zlog, prometheus, and Jaeger. I show a working PoC on dummy code here [1] too. This method is the absolute fastest way to record the minimum info and in tests only spends about 50 nanoseconds per record event.

This means if we wanted to, we can record everything without sampling and it's still much faster than sampling with Jaeger. Why? In tests then just recording in the time critical code is about 50 times fasters than logging, which is already faster than Jaeger. Therefore if Jaeger is sampling every 16th TX, because this other technique is 50 times faster, it's possible to sample all TXs and still be faster. This also means that we can add almost arbitrary extra timing instrumentation and still sample all TXs and still be faster...

[1] https://www.notion.so/dapperlabs/Examine-instrumentation-timing-noodle-on-best-design-9abf3b0ce0c34b7fac6e4dc0c4f67603#e190789d2e33497aaa395f34da11712b