@SaveTheRbtz 2022-07-19
@turbolent on the cadence side perf impact is negligible since it only operates on (already escaping) data structures.  Maybe even slightly faster on the newer version of the code given that it does not have a separate `Time.Now()` call.   That said, let me run cadence specific tests just in case.

<details>
It is quite hard to get a clean run on my laptop and even on a linux server (likely due to Turbo boost), but at least it shows that allocations did not change.
  ```
name                                              old time/op    new time/op    delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:arm64
RuntimeFungibleTokenTransfer-8                       348µs ± 2%     348µs ± 4%    ~     (p=0.541 n=8+9)
RuntimeResourceDictionaryValues-8                   2.35ms ± 3%    2.37ms ± 5%    ~     (p=0.631 n=10+10)
pkg:github.com/onflow/cadence/runtime/interpreter goos:darwin goarch:arm64
ValueIsSubtypeOfSemaType-8                          41.6ns ± 1%    41.5ns ± 2%    ~     (p=0.305 n=9+10)
pkg:github.com/onflow/cadence/runtime/parser goos:darwin goarch:arm64
ParseDeploy/byte_array-8                            6.84ms ± 1%    6.90ms ± 1%  +0.95%  (p=0.000 n=10+9)
ParseDeploy/decode_hex-8                             597µs ± 2%     600µs ± 1%    ~     (p=0.113 n=10+9)
ParseInfix-8                                        4.00µs ± 1%    4.01µs ± 1%    ~     (p=0.085 n=10+10)
ParseFungibleToken/Without_memory_metering-8        89.7µs ± 1%    90.3µs ± 1%    ~     (p=0.065 n=10+9)
ParseFungibleToken/With_memory_metering-8            109µs ± 2%     114µs ± 3%  +3.95%  (p=0.000 n=10+10)
ParseArray-8                                        4.62ms ± 1%    4.73ms ± 2%  +2.43%  (p=0.000 n=9+10)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:arm64
QualifiedIdentifierCreation/One_level-8             2.06ns ± 1%    2.04ns ± 1%  -0.81%  (p=0.001 n=9+9)
QualifiedIdentifierCreation/Three_levels-8          63.2ns ± 2%    62.3ns ± 0%  -1.49%  (p=0.000 n=9+9)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:arm64
ContractInterfaceFungibleToken-8                    21.0µs ± 1%    20.9µs ± 0%  -0.78%  (p=0.001 n=9+10)
CheckContractInterfaceFungibleTokenConformance-8    65.7µs ± 1%    64.8µs ± 3%  -1.49%  (p=0.019 n=9+9)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:arm64
NewInterpreter/new_interpreter-8                     524ns ± 2%     521ns ± 1%    ~     (p=0.055 n=10+10)
NewInterpreter/new_sub-interpreter-8                1.15µs ± 3%    1.13µs ± 1%  -1.44%  (p=0.027 n=10+10)
InterpretRecursionFib-8                             1.64ms ± 3%    1.61ms ± 2%  -2.35%  (p=0.006 n=10+9)

name                                              old alloc/op   new alloc/op   delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:arm64
RuntimeFungibleTokenTransfer-8                       233kB ± 1%     233kB ± 1%    ~     (p=0.549 n=10+9)
RuntimeResourceDictionaryValues-8                   2.23MB ± 0%    2.23MB ± 0%    ~     (p=0.497 n=10+9)
pkg:github.com/onflow/cadence/runtime/interpreter goos:darwin goarch:arm64
ValueIsSubtypeOfSemaType-8                           48.0B ± 0%     48.0B ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/parser goos:darwin goarch:arm64
ParseDeploy/byte_array-8                            4.38MB ± 3%    4.38MB ± 1%    ~     (p=0.853 n=10+10)
ParseDeploy/decode_hex-8                             214kB ± 0%     214kB ± 0%    ~     (p=0.438 n=10+10)
ParseInfix-8                                        2.17kB ± 0%    2.17kB ± 0%    ~     (p=1.000 n=10+10)
ParseFungibleToken/Without_memory_metering-8        36.3kB ± 0%    36.3kB ± 0%    ~     (p=0.641 n=10+10)
ParseFungibleToken/With_memory_metering-8           36.3kB ± 0%    36.3kB ± 0%    ~     (p=0.126 n=8+10)
ParseArray-8                                        2.78MB ± 1%    2.78MB ± 1%    ~     (p=0.853 n=10+10)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:arm64
QualifiedIdentifierCreation/One_level-8              0.00B          0.00B         ~     (all equal)
QualifiedIdentifierCreation/Three_levels-8           64.0B ± 0%     64.0B ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:arm64
ContractInterfaceFungibleToken-8                    25.1kB ± 0%    25.1kB ± 0%  -0.00%  (p=0.003 n=10+8)
CheckContractInterfaceFungibleTokenConformance-8    63.6kB ± 0%    63.6kB ± 0%  -0.00%  (p=0.029 n=10+7)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:arm64
NewInterpreter/new_interpreter-8                      928B ± 0%      928B ± 0%    ~     (all equal)
NewInterpreter/new_sub-interpreter-8                1.46kB ± 0%    1.46kB ± 0%    ~     (all equal)
InterpretRecursionFib-8                             1.51MB ± 0%    1.51MB ± 0%    ~     (p=0.598 n=10+10)

name                                              old allocs/op  new allocs/op  delta
pkg:github.com/onflow/cadence/runtime goos:darwin goarch:arm64
RuntimeFungibleTokenTransfer-8                       4.16k ± 0%     4.16k ± 0%    ~     (all equal)
RuntimeResourceDictionaryValues-8                    37.4k ± 0%     37.4k ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/interpreter goos:darwin goarch:arm64
ValueIsSubtypeOfSemaType-8                            1.00 ± 0%      1.00 ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/parser goos:darwin goarch:arm64
ParseDeploy/byte_array-8                              105k ± 0%      105k ± 0%    ~     (p=0.671 n=10+10)
ParseDeploy/decode_hex-8                              77.0 ± 0%      77.0 ± 0%    ~     (all equal)
ParseInfix-8                                          66.0 ± 0%      66.0 ± 0%    ~     (all equal)
ParseFungibleToken/Without_memory_metering-8         1.06k ± 0%     1.06k ± 0%    ~     (all equal)
ParseFungibleToken/With_memory_metering-8            1.06k ± 0%     1.06k ± 0%    ~     (all equal)
ParseArray-8                                         70.0k ± 0%     70.0k ± 0%    ~     (p=0.838 n=10+10)
pkg:github.com/onflow/cadence/runtime/sema goos:darwin goarch:arm64
QualifiedIdentifierCreation/One_level-8               0.00           0.00         ~     (all equal)
QualifiedIdentifierCreation/Three_levels-8            2.00 ± 0%      2.00 ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/checker goos:darwin goarch:arm64
ContractInterfaceFungibleToken-8                       441 ± 0%       441 ± 0%    ~     (all equal)
CheckContractInterfaceFungibleTokenConformance-8     1.05k ± 0%     1.05k ± 0%    ~     (all equal)
pkg:github.com/onflow/cadence/runtime/tests/interpreter goos:darwin goarch:arm64
NewInterpreter/new_interpreter-8                      13.0 ± 0%      13.0 ± 0%    ~     (all equal)
NewInterpreter/new_sub-interpreter-8                  42.0 ± 0%      42.0 ± 0%    ~     (all equal)
InterpretRecursionFib-8                              33.5k ± 0%     33.5k ± 0%    ~     (all equal)

  ```
  
</details>

PS. Flow tracing microbenchmarks are here: https://github.com/onflow/flow-go/pull/2806 (i'm missing a second reviewer there =))
PPS. I'm of the first school of thought: IMHO if we get into the nano-second territory we should just use pprof (incl. trace).  For everything else we can always tweak sampling probability.  (Tests are there to check that we do not block and maintain a sane amount of allocations.)